# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Bookkeeping Agent API
  description: REST API for the Bookkeeping Agent MVP — transaction management, AI-powered chat, reports, and alerts for small shop owners.
  version: 1.0.0
  contact:
    name: Bookkeeping Agent Team
servers:
  - url: https://api.yourdomain.com/api/v1
    description: Production
  - url: http://localhost:4000/api/v1
    description: Local development
security:
  - BearerAuth: []
...
paths:
...
  /shops/{shopId}/transactions:
    parameters:
      - $ref: '#/components/parameters/shopId'

    get:
      tags: [transactions]
      summary: List transactions with optional filters
      x-rateLimit: '120 requests per minute'
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [sale, expense, receivable, payable]
        - name: category
          in: query
          schema:
            type: string
        - name: dateFrom
          in: query
          schema:
            type: string
            format: date
          example: '2024-01-01'
        - name: dateTo
          in: query
          schema:
            type: string
            format: date
          example: '2024-01-31'
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 200
      responses:
        '200':
          description: Paginated transaction list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      tags: [transactions]
      summary: Create a transaction
      description: Claude categorizes the transaction automatically if `category` is omitted.
      x-rateLimit: '300 requests per minute'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTransactionRequest'
            example:
              type: sale
              amount: 5000
              currency: NGN
              description: 'Sold 10 bags of rice'
              date: '2024-01-15'
              counterparty: 'Chukwu Market'
      responses:
        '201':
          description: Transaction created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /shops/{shopId}/transactions/{transactionId}:
    parameters:
      - $ref: '#/components/parameters/shopId'
      - $ref: '#/components/parameters/transactionId'

    get:
      tags: [transactions]
      summary: Get a single transaction
      x-rateLimit: '120 requests per minute'
      responses:
        '200':
          description: Transaction
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    patch:
      tags: [transactions]
      summary: Update a transaction
      x-rateLimit: '120 requests per minute'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTransactionRequest'
      responses:
        '200':
          description: Updated transaction
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    delete:
      tags: [transactions]
      summary: Delete a transaction
      x-rateLimit: '60 requests per minute'
      responses:
        '204':
          description: Deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  # ── CHAT ─────────────────────────────────────
...
  /reports/receipt:
    post:
      tags: [reports]
      summary: Generate a receipt PDF for a transaction
      x-rateLimit: '30 requests per minute'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [transactionId]
              properties:
                transactionId:
                  type: string
                  format: uuid
                shopId:
                  type: string
                  format: uuid
      responses:
        '200':
          description: PDF file
          content:
            application/pdf:
              schema:
                type: string
                format: binary
          headers:
            Content-Disposition:
              schema:
                type: string
              example: 'attachment; filename="receipt-2024-01-15.pdf"'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /reports/credit:
    post:
      tags: [reports]
      summary: Generate a credit (receivables) report PDF
      x-rateLimit: '10 requests per minute'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DateRangeReportRequest'
      responses:
        '200':
          description: PDF file
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /reports/stock:
    post:
      tags: [reports]
      summary: Generate a stock/inventory report PDF
      x-rateLimit: '10 requests per minute'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DateRangeReportRequest'
      responses:
        '200':
          description: PDF file
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /reports/pl:
    post:
      tags: [reports]
      summary: Generate a Profit & Loss statement PDF
      x-rateLimit: '10 requests per minute'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DateRangeReportRequest'
      responses:
        '200':
          description: PDF file
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

  # ── ALERTS ───────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    shopId:
      name: shopId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    transactionId:
      name: transactionId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Invalid request body or parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: BAD_REQUEST
            message: 'amount must be a positive number'
            details: []
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: UNAUTHORIZED
            message: 'Token is expired or missing'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: NOT_FOUND
            message: 'Shop not found'
...
    InternalError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: INTERNAL_SERVER_ERROR
            message: 'An unexpected error occurred'
  schemas:
...
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          description: Machine-readable error code
        message:
          type: string
          description: Human-readable description
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
...
    Transaction:
      type: object
      properties:
        id:
          type: string
          format: uuid
        shopId:
          type: string
          format: uuid
        type:
          type: string
          enum: [sale, expense, receivable, payable]
        amount:
          type: number
          format: float
        currency:
          type: string
        description:
          type: string
        category:
          type: string
          description: Auto-assigned by Claude if not provided
        counterparty:
          type: string
          description: Customer or supplier name
        date:
          type: string
          format: date
        dueDate:
          type: string
          format: date
          description: For receivables/payables
        status:
          type: string
          enum: [pending, settled, overdue]
        aiCategorized:
          type: boolean
        importId:
          type: string
          format: uuid
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateTransactionRequest:
      type: object
      required: [type, amount, date]
      properties:
        type:
          type: string
          enum: [sale, expense, receivable, payable]
        amount:
          type: number
          minimum: 0.01
        currency:
          type: string
          default: NGN
        description:
          type: string
          maxLength: 500
        category:
          type: string
          description: If omitted, Claude categorizes automatically
        counterparty:
          type: string
        date:
          type: string
          format: date
        dueDate:
          type: string
          format: date
    UpdateTransactionRequest:
      type: object
      properties:
        type:
          type: string
          enum: [sale, expense, receivable, payable]
        amount:
          type: number
          minimum: 0.01
        description:
          type: string
        category:
          type: string
        counterparty:
          type: string
        date:
          type: string
          format: date
        dueDate:
          type: string
          format: date
        status:
          type: string
          enum: [pending, settled, overdue]
    TransactionListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Transaction'
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # ── Chat ────────────────────────────────────
...
    DateRangeReportRequest:
      type: object
      required: [shopId, dateFrom, dateTo]
      properties:
        shopId:
          type: string
          format: uuid
        dateFrom:
          type: string
          format: date
        dateTo:
          type: string
          format: date
        sendEmail:
          type: boolean
          default: false
          description: Also email the report to the authenticated user

    # ── Alerts ───────────────────────────────────
...